+++ packages/react/src/components/logged-out/index.tsx
@@ -7,12 +7,12 @@ type Props = {
 /**
  * Renders the children only when the SDK is ready and the user is not authenticated.
  */
-export const LoggedOut: React.FC<Props> = ({ children }) => {
-  const { user } = useSlashID();
+export const LoggedOut = ({ children }: Props) => {
+  const { isAuthenticated, isLoading } = useSlashID();
 
-  if (user) {
+  if (isAuthenticated || isLoading) {
     return null;
   }
 
-  return <>{children}</>;
-};
+  return (<>{children}</>)
+}