+++ apps/react-nextjs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # react-nextjs
 
+## 0.1.27
+
+### Patch Changes
+
+- Updated dependencies [02ea6e5]
+  - @slashid/react@1.8.0
+  - demo-form@0.0.26
+
 ## 0.1.26
 
 ### Patch Changes